requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-jose==3.5.0
passlib==1.7.4
python-dotenv==1.1.1
sqlalchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.16.2
httpx==0.28.1
email-validator==2.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
